+++ packages/core-js/proposals/using-statement.js
@@ -1,0 +1,2 @@
+// https://github.com/tc39/proposal-using-statement
+require('../modules/esnext.symbol.dispose');